Moved commit model changes to config validation page
This checkin includes:
1. Moved commit model changes to config validation page
2. Ajusted messages accordingly.
3. Removed commit save state
4. Cleaned commit refrence in deploy progress page
