+++ src/localization/bundles/en.json
@@ -83,9 +83,9 @@
     "validate.config.files.deploy": "Deploy",
     "validate.config.files.msg.info1": "Click on a configuration file to view or edit its content.",
     "validate.config.files.msg.info2": "Click the Validate button to run the configuration processor.",
-    "validate.config.files.msg.valid": "The configuration processor completed successfully. Data model is valid.",
-    "validate.config.files.msg.invalid": "The configuration processor failed with the following message: ",
-    "validate.config.files.msg.validating": "Validating...",
+    "validate.config.files.msg.valid": "The configuration processor completed successfully. Data model is valid and changes has been committed.",
+    "validate.config.files.msg.invalid": "Validating and committing data model changes failed with the following message: ",
+    "validate.config.files.msg.validating": "Validating and committing data model changes...",
     "validate.deployment.doWipeDisks": "Wipe Data Disks",
     "validate.deployment.doWipeDisks.tooltip": "Wipes the GPT(MBR) on the disks' partitions before cloud deployment.",
     "validate.deployment.encryptKey": "Encryption Key",
@@ -319,7 +319,6 @@
     "deploy.progress.step5": "Quick health check",
     "deploy.progress.step6": "Additional Setup and Configuration",
     "deploy.progress.step7": "n/a",
-    "deploy.progress.commit": "Commit Input Model",
     "deploy.commit.failure": "Failed to commit Ardana DayZero Installer changes. {0}",
     "deploy.fail.check.playbook": "Failed to check in-progress playbook status for {0} - {1}. {2}",
     "deploy.fail.launch.playbook": "Failed to launch playbook {0}. {1}",
